refactor into functions
