+++ cmd/ltkeycloak/from_mattermost.go
@@ -1,0 +1,280 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+	"runtime"
+	"sync"
+	"time"
+
+	gocloak "github.com/Nerzal/gocloak/v13"
+	"github.com/mattermost/mattermost-load-test-ng/deployment"
+	"github.com/mattermost/mattermost-load-test-ng/deployment/terraform"
+	"github.com/mattermost/mattermost/server/public/model"
+	"github.com/mattermost/mattermost/server/public/shared/mlog"
+	"github.com/spf13/cobra"
+)
+
+func userTxtWriter(usersTxtChan chan string, doneChan chan struct{}, userPassword string) {
+	usersTxtFile, err := os.OpenFile("users.txt", os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		mlog.Error("failed to open users.txt file")
+	}
+	defer usersTxtFile.Close()
+
+	for {
+		select {
+		case email := <-usersTxtChan:
+			_, err := usersTxtFile.Write([]byte(model.UserAuthServiceSaml + ":" + email + " " + userPassword + "\n"))
+			if err != nil {
+				mlog.Error("failed to write to users.txt", mlog.String("err", err.Error()))
+			}
+
+		case <-doneChan:
+			return
+		}
+	}
+}
+
+type workerConfig struct {
+	workerNumber     int
+	usersChan        chan *model.User
+	wg               *sync.WaitGroup
+	mmClient         *model.Client4
+	keycloakClient   *gocloak.GoCloak
+	keycloakToken    *gocloak.JWT
+	keycloakRealm    string
+	userPassword     string
+	deploymentConfig *deployment.Config
+	doneChan         chan struct{}
+}
+
+func migrateMattermostUsersToKeycloak(worker *workerConfig) {
+	refreshTokenTicker := time.NewTicker(30 * time.Second)
+	defer refreshTokenTicker.Stop()
+
+	for {
+		ctx := context.Background()
+
+		select {
+		case user := <-worker.usersChan:
+			kcUserID, err := worker.keycloakClient.CreateUser(ctx, worker.keycloakToken.AccessToken, worker.keycloakRealm, gocloak.User{
+				Username:      &user.Username,
+				Email:         &user.Email,
+				Enabled:       gocloak.BoolP(true),
+				EmailVerified: &user.EmailVerified,
+				Credentials: &[]gocloak.CredentialRepresentation{
+					{
+						Temporary: gocloak.BoolP(false),
+						Type:      gocloak.StringP("password"),
+						Value:     gocloak.StringP(worker.userPassword),
+					},
+				},
+			})
+			if err != nil {
+				// Ignore already existing users
+				if apiErr, ok := err.(*gocloak.APIError); ok && apiErr.Code == 409 {
+					mlog.Debug("user already exists in keycloak", mlog.String("username", user.Username))
+					continue
+				}
+
+				mlog.Error("failed to create user in keycloak", mlog.String("err", err.Error()))
+				continue
+			}
+
+			user.AuthData = model.NewString(kcUserID)
+			user.AuthService = model.UserAuthServiceSaml
+			user.Password = ""
+
+			_, _, err = worker.mmClient.UpdateUser(ctx, user)
+			if err != nil {
+				mlog.Error("failed to update user in mattermost", mlog.String("err", err.Error()))
+				continue
+			}
+			mlog.Info("migrated user", mlog.String("username", user.Username))
+			worker.wg.Done()
+
+		case <-refreshTokenTicker.C:
+			mlog.Info("refreshing keycloak token", mlog.Int("worker", worker.workerNumber))
+			var err error
+			worker.keycloakToken, err = worker.keycloakClient.LoginAdmin(
+				ctx,
+				worker.deploymentConfig.ExternalAuthProviderSettings.KeycloakAdminUser,
+				worker.deploymentConfig.ExternalAuthProviderSettings.KeycloakAdminPassword,
+				"master", // TODO: Allow specifying the master realm
+			)
+			if err != nil {
+				mlog.Error("failed to refresh keycloak token", mlog.String("err", err.Error()))
+				close(worker.doneChan)
+			}
+
+		case <-worker.doneChan:
+			return
+		}
+	}
+}
+
+func RunSyncFromMattermostCommandF(cmd *cobra.Command, _ []string) error {
+	started := time.Now()
+
+	ltConfigPath, err := cmd.Flags().GetString("config")
+	if err != nil {
+		return fmt.Errorf("failed to read flag: %w", err)
+	}
+
+	// Guess where the deployment configuration is located
+	deploymentConfig, err := deployment.ReadConfig(filepath.Join(filepath.Dir(ltConfigPath), "deployer"+filepath.Ext(ltConfigPath)))
+	if err != nil {
+		return fmt.Errorf("failed to read deployment configuration: %w", err)
+	}
+
+	dryRun, err := cmd.Flags().GetBool("dry-run")
+	if err != nil {
+		return fmt.Errorf("failed to read flag: %w", err)
+	}
+
+	mmClient := model.NewAPIv4Client("http://" + deploymentConfig.SiteURL)
+
+	_, _, err = mmClient.Login(cmd.Context(), deploymentConfig.AdminEmail, deploymentConfig.AdminPassword)
+	if err != nil {
+		return fmt.Errorf("failed to login to mattermost: %w", err)
+	}
+
+	var keycloakHost string
+	keycloakHost, err = cmd.Flags().GetString("keycloak-host")
+	if err != nil {
+		return fmt.Errorf("failed to read flag: %w", err)
+	}
+
+	// Use the terraform output terraform host if a manual one is not provided. Useful for development.
+	if keycloakHost == "" {
+		t, err := terraform.New("", *deploymentConfig)
+		if err != nil {
+			return fmt.Errorf("failed to create terraform client: %w", err)
+		}
+		terraformOutput, err := t.Output()
+		if err != nil {
+			return fmt.Errorf("failed to get terraform output: %w", err)
+		}
+		if terraformOutput.KeycloakServer.PublicDNS == "" {
+			return fmt.Errorf("keycloak database cluster not found in terraform output")
+		}
+		keycloakHost = terraformOutput.KeycloakServer.PublicDNS
+	}
+
+	keycloakRealm, err := cmd.Flags().GetString("keycloak-realm")
+	if err != nil {
+		return fmt.Errorf("failed to read flag: %w", err)
+	}
+
+	userPassword, err := cmd.Flags().GetString("set-user-password-to")
+	if err != nil {
+		return fmt.Errorf("failed to read flag: %w", err)
+	}
+
+	keycloakClient := gocloak.NewClient(keycloakHost)
+	ctx := context.Background()
+
+	keycloakToken, err := keycloakClient.LoginAdmin(
+		ctx,
+		deploymentConfig.ExternalAuthProviderSettings.KeycloakAdminUser,
+		deploymentConfig.ExternalAuthProviderSettings.KeycloakAdminPassword,
+		"master", // TODO: Allow specifying the master realm
+	)
+	if err != nil {
+		return fmt.Errorf("failed to refresh keycloak token: %w", err)
+	}
+
+	doneChan := make(chan struct{})
+	workers := runtime.NumCPU()
+	if workers > 4 {
+		workers = 4
+	}
+	usersTxtChan := make(chan string, workers*2)
+	usersChan := make(chan *model.User, workers*2)
+
+	wg := sync.WaitGroup{}
+
+	// Worker to write to the users.txt file
+	go userTxtWriter(usersTxtChan, doneChan, userPassword)
+
+	// Workers to run the sync and call the MM and Keycloak APIs
+	for i := 0; i < workers; i++ {
+		workerNumber := i + 1
+
+		go migrateMattermostUsersToKeycloak(&workerConfig{
+			workerNumber:     workerNumber,
+			usersChan:        usersChan,
+			wg:               &wg,
+			mmClient:         mmClient,
+			keycloakClient:   keycloakClient,
+			keycloakToken:    keycloakToken,
+			keycloakRealm:    keycloakRealm,
+			userPassword:     userPassword,
+			deploymentConfig: deploymentConfig,
+			doneChan:         doneChan,
+		})
+	}
+
+	page := 0
+	perPage := 100
+	for {
+		mlog.Info("fetching mattermost users", mlog.Int("page", page), mlog.Int("per_page", perPage))
+		users, _, err := mmClient.GetUsers(ctx, page, perPage, "")
+		if err != nil {
+			return fmt.Errorf("failed to get users from mattermost: %w", err)
+		}
+
+		for _, user := range users {
+			// Skip bots
+			if user.IsBot {
+				continue
+			}
+
+			// Write the user to the users.txt file
+			usersTxtChan <- user.Email
+
+			// Already migrated
+			if user.AuthService == model.UserAuthServiceSaml {
+				continue
+			}
+
+			if !dryRun {
+				wg.Add(1)
+				usersChan <- user
+			} else {
+				mlog.Info("dry-run: would migrate user", mlog.String("username", user.Username))
+			}
+		}
+
+		if len(users) == 0 {
+			break
+		}
+
+		page++
+	}
+
+	wg.Wait()
+	close(doneChan)
+
+	finished := time.Now()
+
+	mlog.Info("migration finished", mlog.Duration("duration", finished.Sub(started)))
+
+	return nil
+}
+
+func MakeSyncFromMattermostCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "from_mattermost",
+		Short: "Migrate data from Mattermost to Keycloak",
+		RunE:  RunSyncFromMattermostCommandF,
+	}
+
+	cmd.Flags().String("set-user-password-to", "testpassword", "Set's the user password to the provided value")
+	cmd.Flags().String("keycloak-realm", "master", "The Keycloak realm to migrate users to")
+
+	return cmd
+}
